Clean up JSON Schema
We made some changes in the past to inline a bunch of $refs to make
things easier for sdk generation. However, this came with downsides of
less distinct types.

This should have a more compatible version, that still preserves the
non-inlined types.

At least on the TypeScript side this has vastly improved the generated
types.